Fix crashes related to writing live output pgn/json files
The crahes are related to Tournament object accessing ChessGame object
state, but these two objects are actually at different sides of slot
and signal when a game is running. This translates to two different
threads accessing the same state without protection. In practice,
ChessGame may update the move list and board state while Tournament is
reading the state in order to write the live pgn/json files.

Fix this by moving the live file update from Tournament to ChessGame.
This way the state is accessed only from a single thread.